arreglos tipo de dato

+++ filesystem/src/estructurafs.c
@@ -1,6 +1,6 @@
 #include "includes/estructurafs.h"
 
-t_bitarray* cargar_bitmap(char* mount_dir, size_t block_count) {
+t_bitarray* cargar_bitmap(char* mount_dir, uint32_t block_count) {
     char filepath[256];
     snprintf(filepath, sizeof(filepath), "%s/bitmap.dat", mount_dir);
     log_info(log_filesystem, "Cargando bitmap desde %s", filepath);
@@ -8,9 +8,9 @@ t_bitarray* cargar_bitmap(char* mount_dir, size_t block_count) {
     FILE* file = fopen(filepath, "rb");
     struct stat st;
     stat(filepath, &st);
-    size_t expected_size = (size_t)ceil((double)block_count / 8.0);
+    uint32_t expected_size = (uint32_t)ceil((double)block_count / 8.0);
     if (st.st_size != expected_size) {
-        log_error(log_filesystem, "El tamaño del archivo %s no es el esperado. Esperado: %zu, Actual: %zu", filepath, expected_size, st.st_size);
+        log_error(log_filesystem, "El tamaño del archivo %s no es el esperado. Esperado: %u, Actual: %lu", filepath, expected_size, st.st_size);
         fclose(file);
         return NULL;
     }
@@ -26,10 +26,7 @@ t_bitarray* cargar_bitmap(char* mount_dir, size_t block_count) {
 }
 
 
-
-//nombre, tamanio, contenido
-int crear_archivo_dump(t_args_dump_memory* info, t_bitarray* bitmap, const char* mount_dir, int block_size) {
-    
+int crear_archivo_dump(t_args_dump_memory* info, t_bitarray* bitmap, const char* mount_dir, uint32_t block_size) {
     int bloques_necesarios = (info->tamanio_proceso / block_size) + 1;
     // 1. verifico si hay espacio disponible
     if(!hay_espacio_disponible(bitmap, bloques_necesarios + 1)){
@@ -37,7 +34,7 @@ int crear_archivo_dump(t_args_dump_memory* info, t_bitarray* bitmap, const char*
     }
 
     // 2. reservo los bloques necesarios
-    size_t* bloque_reservado = malloc((bloques_necesarios + 1) * sizeof(size_t));
+    uint32_t* bloque_reservado = malloc((bloques_necesarios + 1) * sizeof(uint32_t));
     reservar_bloque(bitmap, bloque_reservado, bloques_necesarios + 1);
 
     // 3. Creo el archivo
@@ -57,7 +54,6 @@ int crear_archivo_dump(t_args_dump_memory* info, t_bitarray* bitmap, const char*
 
     free(bloque_reservado);
     return 1;
-
 }
 
 bool hay_espacio_disponible(t_bitarray* bitmap, int bloques_necesarios) {
@@ -75,7 +71,7 @@ bool hay_espacio_disponible(t_bitarray* bitmap, int bloques_necesarios) {
     return false;
 }
 
-void reservar_bloque(t_bitarray* bitmap, size_t* bloques_reservados, size_t bloques_necesarios) {
+void reservar_bloque(t_bitarray* bitmap, uint32_t* bloques_reservados, uint32_t bloques_necesarios) {
     int contador_reserva = 0;
     for(int i = 0; i < bitarray_get_max_bit(bitmap) && contador_reserva < bloques_necesarios; i++){
         if(!bitarray_test_bit(bitmap, i)) {
@@ -86,7 +82,7 @@ void reservar_bloque(t_bitarray* bitmap, size_t* bloques_reservados, size_t bloq
     }
 }
 
-int crear_archivo_metadata(char* filepath, t_args_dump_memory* info, size_t* bloque_reservados, size_t bloques_necesarios) {
+int crear_archivo_metadata(char* filepath, t_args_dump_memory* info, uint32_t* bloque_reservados, uint32_t bloques_necesarios) {
     FILE* archivo_metadata = fopen(filepath, "w");
     if (archivo_metadata == NULL) {
         log_error(log_filesystem, "Error al crear el archivo de metadata");
@@ -94,8 +90,8 @@ int crear_archivo_metadata(char* filepath, t_args_dump_memory* info, size_t* blo
     }
     log_info(log_filesystem, "TAMANIO=%d\n", info->tamanio_proceso);
     log_info(log_filesystem, "BLOQUES=[");
-    for (size_t i = 1; i < bloques_necesarios; i++) {
-        log_info(log_filesystem, "%zu", bloque_reservados[i]);
+    for (uint32_t i = 1; i < bloques_necesarios; i++) {
+        log_info(log_filesystem, "%u", bloque_reservados[i]);
         if (i < bloques_necesarios - 1) {
             log_info(log_filesystem, ",");
         }
@@ -105,7 +101,7 @@ int crear_archivo_metadata(char* filepath, t_args_dump_memory* info, size_t* blo
     return 0;
 }
 
-int escribir_bloques(const char* mount_dir, size_t* bloque_reservados, size_t bloques_necesarios, t_args_dump_memory* info, int block_size) {
+int escribir_bloques(const char* mount_dir, uint32_t* bloque_reservados, uint32_t bloques_necesarios, t_args_dump_memory* info, int block_size) {
     int retardo = config_get_int_value(config, "RETARDO_ACCESO_BLOQUE");
 
     char bloques_filepath[256];
@@ -116,11 +112,11 @@ int escribir_bloques(const char* mount_dir, size_t* bloque_reservados, size_t bl
         return -1;
     }
 
-    size_t bytes_written = 0;
-    for (size_t i = 0; i < bloques_necesarios; i++) {
-        size_t block_index = bloque_reservados[i];
+    uint32_t bytes_written = 0;
+    for (uint32_t i = 0; i < bloques_necesarios; i++) {
+        uint32_t block_index = bloque_reservados[i];
         off_t offset = block_index * block_size;
-        size_t bytes_to_write = (info->tamanio_proceso - bytes_written > block_size) ? block_size : info->tamanio_proceso - bytes_written;
+        uint32_t bytes_to_write = (info->tamanio_proceso - bytes_written > block_size) ? block_size : info->tamanio_proceso - bytes_written;
         if (pwrite(bloques_fd, info->contenido + bytes_written, bytes_to_write, offset) != bytes_to_write) {
             log_error(log_filesystem, "Error al escribir en el bloque");
             close(bloques_fd);
